+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/IssuerAgent.kt
@@ -3,39 +3,24 @@ package at.asitplus.wallet.lib.agent
 import at.asitplus.KmmResult
 import at.asitplus.catching
 import at.asitplus.iso.DeviceKeyInfo
-import at.asitplus.wallet.lib.iso.IssuerSigned
-import at.asitplus.wallet.lib.iso.MobileSecurityObject
 import at.asitplus.iso.ValidityInfo
 import at.asitplus.iso.ValueDigest
 import at.asitplus.iso.ValueDigestList
 import at.asitplus.signum.indispensable.SignatureAlgorithm
-import at.asitplus.signum.indispensable.cosef.CoseHeader
-import at.asitplus.signum.indispensable.cosef.CoseSigned
 import at.asitplus.signum.indispensable.cosef.toCoseKey
 import at.asitplus.signum.indispensable.josef.ConfirmationClaim
 import at.asitplus.signum.indispensable.josef.toJsonWebKey
 import at.asitplus.wallet.lib.DefaultZlibService
 import at.asitplus.wallet.lib.ZlibService
 import at.asitplus.wallet.lib.agent.SdJwtCreator.toSdJsonObject
-import at.asitplus.wallet.lib.cbor.CoseHeaderCertificate
-import at.asitplus.wallet.lib.cbor.CoseHeaderKeyId
-import at.asitplus.wallet.lib.cbor.CoseHeaderNone
-import at.asitplus.wallet.lib.cbor.SignCose
-import at.asitplus.wallet.lib.cbor.SignCoseFun
+import at.asitplus.wallet.lib.cbor.*
 import at.asitplus.wallet.lib.data.*
-import at.asitplus.wallet.lib.data.VerifiableCredentialJws
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.*
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.MediaTypes
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListInfo
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListTokenPayload
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.agents.communication.primitives.StatusListTokenMediaType
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.PositiveDuration
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.data.rfc3986.UniformResourceIdentifier
-import at.asitplus.wallet.lib.jws.SignJwt
-import at.asitplus.wallet.lib.jws.SignJwtFun
-import at.asitplus.wallet.lib.jws.JwsContentTypeConstants
-import at.asitplus.wallet.lib.jws.JwsHeaderCertOrJwk
-import at.asitplus.wallet.lib.jws.JwsHeaderKeyId
+import at.asitplus.wallet.lib.iso.IssuerSigned
+import at.asitplus.wallet.lib.iso.MobileSecurityObject
+import at.asitplus.wallet.lib.jws.*
 import com.benasher44.uuid.uuid4
 import io.github.aakira.napier.Napier
 import kotlinx.datetime.Clock
@@ -48,32 +33,50 @@ import kotlin.time.Duration
 import kotlin.time.Duration.Companion.hours
 
 /**
- * An agent that only implements [Issuer], i.e. it issues credentials for other agents.
+ * An agent that implements [Issuer], i.e. it issues credentials for other agents.
+ *
+ * For backwards compatibility, this also implements [StatusListIssuer], but this should be separated.
  */
-
 class IssuerAgent(
     override val keyMaterial: KeyMaterial = EphemeralKeyWithoutCert(),
+    @Deprecated("Removed, see StatusListAgent")
     private val validator: Validator = Validator(),
     private val issuerCredentialStore: IssuerCredentialStore = InMemoryIssuerCredentialStore(),
     private val statusListBaseUrl: String = "https://wallet.a-sit.at/backend/credentials/status",
+    @Deprecated("Removed, see StatusListAgent")
     private val statusListAggregationUrl: String? = null,
+    @Deprecated("Removed, see StatusListAgent")
     private val zlibService: ZlibService = DefaultZlibService(),
+    @Deprecated("Removed, see StatusListAgent")
     private val revocationListLifetime: Duration = 48.hours,
     private val clock: Clock = Clock.System,
     override val cryptoAlgorithms: Set<SignatureAlgorithm> = setOf(keyMaterial.signatureAlgorithm),
     private val timePeriodProvider: TimePeriodProvider = FixedTimePeriodProvider,
-    /**
-     * The identifier used in `issuer` properties of issued credentials.
-     * Note that for SD-JWT VC this must be a URI. */
+    /** The identifier used in `issuer` properties of issued credentials. Note that for SD-JWT VC this must be a URI. */
     private val identifier: String = keyMaterial.identifier,
     private val signIssuedSdJwt: SignJwtFun<JsonObject> = SignJwt(keyMaterial, JwsHeaderCertOrJwk()),
     private val signIssuedVc: SignJwtFun<VerifiableCredentialJws> = SignJwt(keyMaterial, JwsHeaderKeyId()),
+    @Deprecated("Removed, see StatusListAgent")
     private val signStatusListJwt: SignJwtFun<StatusListTokenPayload> = SignJwt(keyMaterial, JwsHeaderCertOrJwk()),
     private val signMobileSecurityObject: SignCoseFun<MobileSecurityObject> =
         SignCose(keyMaterial, CoseHeaderNone(), CoseHeaderCertificate()),
+    @Deprecated("Removed, see StatusListAgent")
     private val signStatusListCwt: SignCoseFun<StatusListTokenPayload> =
         SignCose(keyMaterial, CoseHeaderKeyId(), CoseHeaderCertificate()),
-) : Issuer {
+) : Issuer,
+    StatusListIssuer by StatusListAgent(
+        keyMaterial = keyMaterial,
+        validator = validator,
+        issuerCredentialStore = issuerCredentialStore,
+        statusListBaseUrl = statusListBaseUrl,
+        statusListAggregationUrl = statusListAggregationUrl,
+        zlibService = zlibService,
+        revocationListLifetime = revocationListLifetime,
+        clock = clock,
+        timePeriodProvider = timePeriodProvider,
+        signStatusListJwt = signStatusListJwt,
+        signStatusListCwt = signStatusListCwt,
+    ) {
 
     /**
      * Wraps the credential-to-be-issued in [credential] into a single instance of [CredentialToBeIssued],
@@ -256,136 +259,6 @@ class IssuerAgent(
         return Issuer.IssuedCredential.VcSdJwt(vcInSdJwt, credential.scheme)
     }
 
-    /**
-     * Wraps the revocation information from [issuerCredentialStore] into a Status List Token,
-     * returns a JWS representation of that.
-     */
-    override suspend fun issueStatusListJwt(time: Instant?) =
-        signStatusListJwt(
-            MediaTypes.STATUSLIST_JWT,
-            buildStatusListTokenPayload(time.toTimePeriod()),
-            StatusListTokenPayload.serializer(),
-        ).getOrElse {
-            Napier.w("issueStatusListJwt error", it)
-            throw IllegalStateException("Status token could not be created.", it)
-        }
-
-    /**
-     * Wraps the revocation information from [issuerCredentialStore] into a Status List Token,
-     * returns a CWS representation of that.
-     */
-    override suspend fun issueStatusListCwt(time: Instant?) =
-        issueStatusListCwt(time.toTimePeriod())
-            ?: throw IllegalStateException("Status token could not be created.")
-
-    suspend fun issueStatusListCwt(timePeriod: Int?): CoseSigned<StatusListTokenPayload>? =
-        wrapStatusListTokenInCoseSigned(buildStatusListTokenPayload(timePeriod))
-
-    /**
-     * Wraps the revocation information from [issuerCredentialStore] into a Token Payload
-     */
-    private fun buildStatusListTokenPayload(timePeriod: Int?): StatusListTokenPayload =
-        StatusListTokenPayload(
-            statusList = buildStatusList(timePeriod),
-            issuedAt = clock.now(),
-            timeToLive = PositiveDuration(revocationListLifetime),
-            subject = UniformResourceIdentifier(
-                getRevocationListUrlFor(timePeriod ?: timePeriodProvider.getCurrentTimePeriod(clock))
-            ),
-        ).also {
-            Napier.d("revocation status list: ${it.statusList}")
-        }
-
-    /**
-     * Returns a status list, where the entry at "revocationListIndex" (of the credential) is INVALID if it is revoked
-     */
-    override fun buildStatusList(timePeriod: Int?): StatusList =
-        StatusList(
-            view = buildStatusListView(timePeriod),
-            aggregationUri = statusListAggregationUrl,
-            zlibService = zlibService,
-        )
-
-    private fun buildStatusListView(timePeriod: Int?): StatusListView =
-        issuerCredentialStore.getStatusListView(timePeriod ?: timePeriodProvider.getCurrentTimePeriod(clock))
-
-    /**
-     * Revokes all verifiable credentials from [credentialsToRevoke] list that parse and validate.
-     * It returns true if all revocations was successful.
-     */
-    override suspend fun revokeCredentials(credentialsToRevoke: List<String>): Boolean =
-        credentialsToRevoke.map {
-            validator.verifyVcJws(it, null)
-        }.filterIsInstance<Verifier.VerifyCredentialResult.SuccessJwt>().all {
-            issuerCredentialStore.setStatus(
-                vcId = it.jws.vc.id,
-                status = TokenStatus.Invalid,
-                timePeriod = timePeriodProvider.getTimePeriodFor(it.jws.vc.issuanceDate)
-            )
-        }
-
-    /**
-     * Revokes all verifiable credentials with ids from [credentialIdsToRevoke]
-     * It returns true if all revocations was successful.
-     */
-    override fun revokeCredentialsWithId(credentialIdsToRevoke: Map<String, Instant>): Boolean =
-        credentialIdsToRevoke.all {
-            issuerCredentialStore.setStatus(
-                vcId = it.key,
-                status = TokenStatus.Invalid,
-                timePeriod = timePeriodProvider.getTimePeriodFor(it.value),
-            )
-        }
-
-    override suspend fun provideStatusListToken(
-        acceptedContentTypes: List<StatusListTokenMediaType>,
-        time: Instant?,
-    ): Pair<StatusListTokenMediaType, StatusListToken> {
-        val preferedType = acceptedContentTypes.firstOrNull()
-            ?: throw IllegalArgumentException("Argument `acceptedContentTypes` must contain at least one item.")
-
-        return preferedType to when (preferedType) {
-            StatusListTokenMediaType.Jwt -> StatusListToken.StatusListJwt(
-                issueStatusListJwt(time),
-                resolvedAt = clock.now(),
-            )
-
-            StatusListTokenMediaType.Cwt -> StatusListToken.StatusListCwt(
-                issueStatusListCwt(time),
-                resolvedAt = clock.now(),
-            )
-        }
-    }
-
-    override suspend fun provideStatusListAggregation() = StatusListAggregation(
-        statusLists = compileCurrentRevocationLists().map {
-            UniformResourceIdentifier(it)
-        }
-    )
-
-    private fun compileCurrentRevocationLists(): List<String> {
-        val list = mutableListOf<String>()
-        for (timePeriod in timePeriodProvider.getRelevantTimePeriods(clock)) {
-            if (timePeriodProvider.getCurrentTimePeriod(clock) == timePeriod
-                || issuerCredentialStore.getStatusListView(timePeriod).isNotEmpty()
-            ) {
-                list.add(getRevocationListUrlFor(timePeriod))
-            }
-        }
-        return list
-    }
-
-    private suspend fun wrapStatusListTokenInCoseSigned(statusListTokenPayload: StatusListTokenPayload): CoseSigned<StatusListTokenPayload>? =
-        signStatusListCwt(
-            protectedHeader = CoseHeader(type = MediaTypes.Application.STATUSLIST_CWT),
-            unprotectedHeader = null,
-            payload = statusListTokenPayload,
-            serializer = StatusListTokenPayload.serializer(),
-        ).getOrElse {
-            Napier.w("Could not wrapStatusListInJws", it)
-            return null
-        }
-
     private fun getRevocationListUrlFor(timePeriod: Int) = statusListBaseUrl.let {
         it + (if (!it.endsWith('/')) "/" else "") + timePeriod
     }
@@ -399,7 +272,4 @@ class IssuerAgent(
         jwtId = id
     )
 
-    fun Instant?.toTimePeriod() = this?.let {
-        timePeriodProvider.getTimePeriodFor(it)
-    } ?: timePeriodProvider.getCurrentTimePeriod(clock)
 }